requirements.txt
crewai>=0.1.0
crewai-tools>=0.1.0
langchain>=0.1.0
python-dotenv>=1.0.0
gitpython>=3.1.40
pylint>=3.0.0
ruff>=0.1.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
radon>=6.0.0